vps/requirements.txt
fastapi
joblib
python-dotenv
uvicorn
httpx
numpy==1.26.2
joblib==1.3.2